requirements.txt
langchain>=0.1.0
openai>=1.0.0
pymongo>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
pytest>=7.0.0

